Chain_Of_Thought/ddx-cot-poc/requirements.txt
langgraph>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain>=0.1.0
langfuse>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
